pyproject.toml
[project]
name = "freeform-rpg"
version = "0.1.0"
description = "AI-driven narrative RPG with strong continuity mechanics"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.18.0",
    "pyyaml>=6.0",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
rag = [
    "chromadb>=0.4.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
freeform-rpg = "src.cli.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
